feat: fix lint and build

+++ app/api/adapters/db/schema.ts
@@ -4,7 +4,7 @@ import {
   text,
   timestamp,
   varchar,
-  numeric,
+  integer,
 } from "drizzle-orm/pg-core";
 
 const timestamps = {
@@ -29,6 +29,6 @@ export const calendarEventsTable = pgTable("calendar_events", {
   endDateTime: timestamp({ mode: "date" }).notNull(),
   isAllDay: boolean().notNull().default(false),
   color: varchar({ length: 5 }),
-  durationInMinutes: numeric().notNull(),
+  durationInMinutes: integer().notNull(),
   ...timestamps,
 });
